Add a bunch of CGFloat casts

+++ components/TextFields/examples/experimental/supplemental/MDCBaseTextArea+MaterialTheming.m
@@ -230,7 +230,7 @@
 @implementation TextAreaFilledPositioningDelegate
 
 - (CGFloat)assistiveLabelPaddingWithContainerHeight:(CGFloat)containerHeight {
-  return 0.13 * containerHeight;
+  return (CGFloat)0.13 * containerHeight;
 }
 
 - (CGFloat)defaultContainerHeightWithTextHeight:(CGFloat)textHeight {
@@ -250,8 +250,8 @@
                   preferredContainerHeight:(CGFloat)preferredContainerHeight {
   CGFloat containerHeight = [self containerHeightWithTextHeight:textHeight
                                        preferredContainerHeight:preferredContainerHeight];
-  CGFloat offset = containerHeight * 0.20;
-  return offset - (0.5 * floatingLabelHeight);
+  CGFloat offset = containerHeight * (CGFloat)0.20;
+  return offset - ((CGFloat)0.5 * floatingLabelHeight);
 }
 
 - (CGFloat)textMinYWithFloatingLabelWithTextHeight:(CGFloat)textHeight
@@ -259,8 +259,8 @@
                           preferredContainerHeight:(CGFloat)preferredContainerHeight {
   CGFloat containerHeight = [self containerHeightWithTextHeight:textHeight
                                        preferredContainerHeight:preferredContainerHeight];
-  CGFloat offset = containerHeight * 0.64;
-  return offset - (0.5 * textHeight);
+  CGFloat offset = containerHeight * (CGFloat)0.64;
+  return offset - ((CGFloat)0.5 * textHeight);
 }
 
 - (CGFloat)textMinYWithoutFloatingLabelWithTextHeight:(CGFloat)textHeight
@@ -268,8 +268,8 @@
                              preferredContainerHeight:(CGFloat)preferredContainerHeight {
   CGFloat containerHeight = [self containerHeightWithTextHeight:textHeight
                                        preferredContainerHeight:preferredContainerHeight];
-  CGFloat offset = containerHeight * 0.5;
-  return offset - (0.5 * textHeight);
+  CGFloat offset = containerHeight * (CGFloat)0.5;
+  return offset - ((CGFloat)0.5 * textHeight);
 }
 
 @end
@@ -277,7 +277,7 @@
 @implementation TextAreaOutlinedPositioningDelegate
 
 - (CGFloat)assistiveLabelPaddingWithContainerHeight:(CGFloat)containerHeight {
-  return 0.13 * containerHeight;
+  return (CGFloat)0.13 * containerHeight;
 }
 
 - (CGFloat)defaultContainerHeightWithTextHeight:(CGFloat)textHeight {
@@ -295,7 +295,7 @@
 - (CGFloat)floatingLabelMinYWithTextHeight:(CGFloat)textHeight
                        floatingLabelHeight:(CGFloat)floatingLabelHeight
                   preferredContainerHeight:(CGFloat)preferredContainerHeight {
-  return 0 - (0.5 * floatingLabelHeight);
+  return 0 - ((CGFloat)0.5 * floatingLabelHeight);
 }
 
 - (CGFloat)textMinYWithFloatingLabelWithTextHeight:(CGFloat)textHeight
@@ -303,8 +303,8 @@
                           preferredContainerHeight:(CGFloat)preferredContainerHeight {
   CGFloat containerHeight = [self containerHeightWithTextHeight:textHeight
                                        preferredContainerHeight:preferredContainerHeight];
-  CGFloat offset = containerHeight * 0.5;
-  return offset - (0.5 * textHeight);
+  CGFloat offset = containerHeight * (CGFloat)0.5;
+  return offset - ((CGFloat)0.5 * textHeight);
 }
 
 - (CGFloat)textMinYWithoutFloatingLabelWithTextHeight:(CGFloat)textHeight
@@ -312,8 +312,8 @@
                              preferredContainerHeight:(CGFloat)preferredContainerHeight {
   CGFloat containerHeight = [self containerHeightWithTextHeight:textHeight
                                        preferredContainerHeight:preferredContainerHeight];
-  CGFloat offset = containerHeight * 0.5;
-  return offset - (0.5 * textHeight);
+  CGFloat offset = containerHeight * (CGFloat)0.5;
+  return offset - ((CGFloat)0.5 * textHeight);
 }
 
 @end
